refactor: add dedicated module for gas counter tests
